requirements.txt
streamlit==1.32.0
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.4.0
numpy==1.26.4
pandas==2.2.0
openai==1.6.1
geopy==2.4.1
pysolar==0.10
httpx==0.24.1
# torch==2.2.0
# torchvision==0.17.0
# segmentation-models-pytorch==0.3.3
opencv-python==4.9.0.80
matplotlib==3.8.4 